Just starting to break code into main/suppl. files.
I realized that there will be a decent amount of analysis that is not
relevant to the final results of the manuscript, and that anyone looking
to reproduce the analyses may be annoyed at having to differentiate
between the 'real' code and the supporting code. So this commit is just
after I started working on an Rmarkdown document for the supporting
code, but before breaking out the dataprep and main analysis .R files.